Remove unused product images, enhance product and provider management, and add client management functionality

+++ php/delete.php
@@ -5,11 +5,18 @@ require_once("conexion.php");
 if ($data = json_decode(file_get_contents("php://input"), true)) {
     $mysql_query = "DELETE FROM " . $data['table'] . " WHERE " . $data['where'];
 
-    $ruta = "SELECT imagen FROM productos WHERE idProducto = " . $data['id'];
+    if ($data['table'] === "productos") {
+        $sqlruta = "SELECT imagen FROM productos WHERE idProducto = " . $data['id'];
+        $resultado = mysqli_query($conn, $sqlruta);
+        $row = mysqli_fetch_assoc($resultado);
+        $ruta = $row['imagen'];
+    }
 
     if (mysqli_query($conn, $mysql_query)) {
-        deleteImage($ruta);
         echo "Eliminado correctamente";
+        if ($data['table'] === "productos") {
+            deleteImage($ruta);
+        }
     } else {
         echo "Error: " . mysqli_error($conn);
     }
@@ -22,6 +29,11 @@ function deleteImage($ruta)
     $rutaCompleta = "C:/xampp/htdocs/PFSM4/" . $ruta;
 
     if (file_exists($rutaCompleta)) {
-        unlink($rutaCompleta);
+        if (unlink($rutaCompleta)) {
+            echo " ✅";
+        } else {
+            $error = error_get_last();
+            echo "Error al eliminar: " . $error["message"];
+        }
     }
 }
